+++ DIII4A/idTech4Amm/src/main/java/com/karin/idTech4Amm/sys/Constants.java
@@ -12,25 +12,23 @@ public final class Constants
     public static final String CONST_PREFERENCE_APP_CRASH_INFO = "_APP_CRASH_INFO";
     public static final String CONST_PREFERENCE_EXCEPTION_DEBUG = "_EXCEPTION_DEBUG";
 
-    public static final int CONST_UPDATE_RELEASE = 27;
-    public static final String CONST_RELEASE = "2023-04-05";
+    public static final int CONST_UPDATE_RELEASE = 28;
+    public static final String CONST_RELEASE = "2023-04-13";
     public static final String CONST_EMAIL = "beyondk2000@gmail.com";
     public static final String CONST_DEV = "Karin";
     public static final String CONST_CODE = "Harmattan";
     public static final String CONST_APP_NAME = "idTech4A++"; // "DIII4A++";
     public static final String CONST_NAME = "DOOM III/Quake 4/Prey(2006) for Android(Harmattan Edition)";
 	public static final String CONST_MAIN_PAGE = "https://github.com/glKarin/com.n0n3m4.diii4a";
+    public static final String CONST_TIEBA = "https://tieba.baidu.com/p/6825594793";
 	public static final String CONST_DEVELOPER = "https://github.com/glKarin";
     public static final String CONST_DEVELOPER_XDA = "https://forum.xda-developers.com/member.php?u=10584229";
 	public static final String CONST_CHECK_FOR_UPDATE_URL = "https://raw.githubusercontent.com/glKarin/com.n0n3m4.diii4a/package/CHECK_FOR_UPDATE.json";
 	public static final String[] CONST_CHANGES = {
-			"Fixup some line effects in Quake 4. e.g. monster body disappear effect, lines net.",
-            "Fixup radio icon in player HUD right-top corner in Quake 4.",
-            "Fixup dialog interactive in Quake 4. e.g. dialog of creating MP game server.",
-            "Fixup MP game loading GUI and MP game main menu in Quake 4.",
-            "Add integer cvar named `harm_si_autoFillBots` for automatic fill bots after map loaded in MP game in Quake 4(0 to disable). `fillbots` command support append bot num argument.",
-            "Add automatic set bot's team in MP team game, random bot model, and fixup some bot's bugs in Quake 4.",
-            "Add `SABot`'s aas file pak of MP game maps in `Quake 4 helper dialog`.",
+            "Add bool cvar `harm_g_mutePlayerFootStep` to control mute player footstep sound(default on) in Quake 4.",
+            "Fix some light's brightness depend on sound amplitude in Quake 4. e.g. in most levels like `airdefense2`, at some dark passages, it should has a repeat-flashing lighting.",
+            "Remove Quake 4 helper dialog when start Quake 4, if want to extract resource files, open `Other` -> `Extract resource` in menu.",
+            "(Bug)In Quake 4, if load some levels has noise with effects on, typed `bse_enable` to 0, and then typed `bse_enable` back to 1 in console, noise can resolved.",
 	};
     public static final String CONST_PACKAGE = "com.karin.idTech4Amm";
 
@@ -48,7 +46,7 @@ public final class Constants
         public static final String VOLUME_UP_KEY = "harm_volume_up_key";
         public static final String VOLUME_DOWN_KEY = "harm_volume_down_key";
         public static final String HIDE_AD_BAR = "harm_hide_ad_bar";
-		public static final String OPEN_QUAKE4_HELPER = "harm_open_quake4_helper";
+		//public static final String OPEN_QUAKE4_HELPER = "harm_open_quake4_helper";
 		public static final String CONTROLS_CONFIG_POSITION_UNIT = "harm_controls_config_position_unit";
         public static final String COVER_EDGES = "harm_cover_edges";