+++ CHANGES.md
@@ -1,8 +1,18 @@
 ## Change logs
-**1.1.0harmattan27(natasha)** - **2023-04-05**
+**1.1.0harmattan28(natasha)** - **2023-04-13**
 
 ----------------------------------------------------------------------------------
 ### Update
+
+> 1.1.0harmattan28 (2023-04-13)
+
+* Add bool cvar `harm_g_mutePlayerFootStep` to control mute player footstep sound(default on) in Quake 4.
+* Fix some light's brightness depend on sound amplitude in Quake 4. e.g. in most levels like `airdefense2`, at some dark passages, it should has a repeat-flashing lighting.
+* Remove Quake 4 helper dialog when start Quake 4, if want to extract resource files, open `Other` -> `Extract resource` in menu.
+* (Bug)In Quake 4, if load some levels has noise with effects on, typed `bse_enable` to 0, and then typed `bse_enable` back to 1 in console, noise can resolved.
+
+----------------------------------------------------------------------------------
+### History
 
 > 1.1.0harmattan27 (2023-04-05)
 
@@ -15,7 +25,6 @@
 * Add `SABot`'s aas file pak of MP game maps in `Quake 4 helper dialog`.
 
 ----------------------------------------------------------------------------------
-### History
 
 > 1.1.0harmattan26 (2023-03-25)